Refactored the package and groupId to io.magentys instead of com.magentys
