+++ pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.magentys.donut</groupId>
+    <groupId>io.magentys.donut</groupId>
     <artifactId>donut-junit-adapter</artifactId>
     <packaging>jar</packaging>
     <version>0.1-SNAPSHOT</version>
@@ -81,7 +81,7 @@
                     </descriptorRefs>
                     <archive>
                         <manifest>
-                            <mainClass>com.magentys.donut.main.Main</mainClass>
+                            <mainClass>io.magentys.donut.main.Main</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
