--- src/main/java/com/magentys/donut/junit/model/Error.java
@@ -1,107 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.11 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2017.09.22 at 01:07:35 PM AWST 
-//
-
-
-package com.magentys.donut.junit.model;
-
-import javax.xml.bind.annotation.*;
-import java.io.Serializable;
-
-
-/**
- * <p>Java class for anonymous complex type.
- * <p>
- * <p>The following schema fragment specifies the expected content contained within this class.
- * <p>
- * <pre>
- * &lt;complexType&gt;
- *   &lt;complexContent&gt;
- *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
- *       &lt;attribute name="type" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
- *       &lt;attribute name="message" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
- *     &lt;/restriction&gt;
- *   &lt;/complexContent&gt;
- * &lt;/complexType&gt;
- * </pre>
- */
-@XmlAccessorType(XmlAccessType.FIELD)
-@XmlType(name = "", propOrder = {
-        "content"
-})
-@XmlRootElement(name = "error")
-public class Error
-        implements Serializable {
-
-    private final static long serialVersionUID = 1L;
-    @XmlValue
-    protected String content;
-    @XmlAttribute(name = "type")
-    protected String type;
-    @XmlAttribute(name = "message")
-    protected String message;
-
-    /**
-     * Gets the value of the content property.
-     *
-     * @return possible object is
-     * {@link String }
-     */
-    public String getContent() {
-        return content;
-    }
-
-    /**
-     * Sets the value of the content property.
-     *
-     * @param value allowed object is
-     *              {@link String }
-     */
-    public void setContent(String value) {
-        this.content = value;
-    }
-
-    /**
-     * Gets the value of the type property.
-     *
-     * @return possible object is
-     * {@link String }
-     */
-    public String getType() {
-        return type;
-    }
-
-    /**
-     * Sets the value of the type property.
-     *
-     * @param value allowed object is
-     *              {@link String }
-     */
-    public void setType(String value) {
-        this.type = value;
-    }
-
-    /**
-     * Gets the value of the message property.
-     *
-     * @return possible object is
-     * {@link String }
-     */
-    public String getMessage() {
-        return message;
-    }
-
-    /**
-     * Sets the value of the message property.
-     *
-     * @param value allowed object is
-     *              {@link String }
-     */
-    public void setMessage(String value) {
-        this.message = value;
-    }
-
-}
