--- src/main/java/com/magentys/donut/junit/model/ObjectFactory.java
@@ -1,116 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.11 
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
-// Any modifications to this file will be lost upon recompilation of the source schema. 
-// Generated on: 2017.09.22 at 01:07:35 PM AWST 
-//
-
-
-package com.magentys.donut.junit.model;
-
-import javax.xml.bind.JAXBElement;
-import javax.xml.bind.annotation.XmlElementDecl;
-import javax.xml.bind.annotation.XmlRegistry;
-import javax.xml.namespace.QName;
-
-
-/**
- * This object contains factory methods for each
- * Java content interface and Java element interface
- * generated in the generated package.
- * <p>An ObjectFactory allows you to programatically
- * construct new instances of the Java representation
- * for XML content. The Java representation of XML
- * content can consist of schema derived interfaces
- * and classes representing the binding of schema
- * type definitions, element declarations and model
- * groups.  Factory methods for each of these are
- * provided in this class.
- */
-@XmlRegistry
-public class ObjectFactory {
-
-    private final static QName _Skipped_QNAME = new QName("", "skipped");
-    private final static QName _SystemErr_QNAME = new QName("", "system-err");
-    private final static QName _SystemOut_QNAME = new QName("", "system-out");
-
-    /**
-     * Create a new ObjectFactory that can be used to create new instances of schema derived classes for package: generated
-     */
-    public ObjectFactory() {
-    }
-
-    /**
-     * Create an instance of {@link Failure }
-     */
-    public Failure createFailure() {
-        return new Failure();
-    }
-
-    /**
-     * Create an instance of {@link Error }
-     */
-    public Error createError() {
-        return new Error();
-    }
-
-    /**
-     * Create an instance of {@link Properties }
-     */
-    public Properties createProperties() {
-        return new Properties();
-    }
-
-    /**
-     * Create an instance of {@link Property }
-     */
-    public Property createProperty() {
-        return new Property();
-    }
-
-    /**
-     * Create an instance of {@link Testcase }
-     */
-    public Testcase createTestcase() {
-        return new Testcase();
-    }
-
-    /**
-     * Create an instance of {@link Testsuite }
-     */
-    public Testsuite createTestsuite() {
-        return new Testsuite();
-    }
-
-    /**
-     * Create an instance of {@link Testsuites }
-     */
-    public Testsuites createTestsuites() {
-        return new Testsuites();
-    }
-
-    /**
-     * Create an instance of {@link JAXBElement }{@code <}{@link String }{@code >}}
-     */
-    @XmlElementDecl(namespace = "", name = "skipped")
-    public JAXBElement<String> createSkipped(String value) {
-        return new JAXBElement<String>(_Skipped_QNAME, String.class, null, value);
-    }
-
-    /**
-     * Create an instance of {@link JAXBElement }{@code <}{@link String }{@code >}}
-     */
-    @XmlElementDecl(namespace = "", name = "system-err")
-    public JAXBElement<String> createSystemErr(String value) {
-        return new JAXBElement<String>(_SystemErr_QNAME, String.class, null, value);
-    }
-
-    /**
-     * Create an instance of {@link JAXBElement }{@code <}{@link String }{@code >}}
-     */
-    @XmlElementDecl(namespace = "", name = "system-out")
-    public JAXBElement<String> createSystemOut(String value) {
-        return new JAXBElement<String>(_SystemOut_QNAME, String.class, null, value);
-    }
-
-}
