--- src/main/java/com/magentys/donut/reporter/EntityGsonBuilder.java
@@ -1,12 +1,0 @@
-package com.magentys.donut.reporter;
-
-
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-
-class EntityGsonBuilder {
-    static Gson createGson() {
-        return new GsonBuilder().setPrettyPrinting()
-                .create();
-    }
-}
