+++ src/main/java/io/magentys/donut/gherkin/model/Match.java
@@ -1,0 +1,25 @@
+
+package io.magentys.donut.gherkin.model;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+public class Match {
+
+    @SerializedName("location")
+    @Expose
+    private String location;
+
+    public Match(String location) {
+        this.location = location;
+    }
+
+    public String getLocation() {
+        return location;
+    }
+
+    public void setLocation(String location) {
+        this.location = location;
+    }
+
+}
