+++ src/main/java/io/magentys/donut/gherkin/model/Result.java
@@ -1,0 +1,51 @@
+
+package io.magentys.donut.gherkin.model;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+public class Result {
+
+    @SerializedName("duration")
+    @Expose
+    private Long duration;
+
+    @SerializedName("status")
+    @Expose
+    private String status;
+
+    @SerializedName("error_message")
+    @Expose
+    private String errorMessage;
+
+
+    public Result(String status, long duration,String errorMessage) {
+        this.status = status;
+        this.duration = duration;
+        this.errorMessage = errorMessage;
+    }
+
+    public Long getDuration() {
+        return duration;
+    }
+
+    public void setDuration(Long duration) {
+        this.duration = duration;
+    }
+
+    public String getStatus() {
+        return status;
+    }
+
+    public void setStatus(String status) {
+        this.status = status;
+    }
+
+    public String getErrorMessage() {
+        return errorMessage;
+    }
+
+    public void setErrorMessage(String errorMessage) {
+        this.errorMessage = errorMessage;
+    }
+}
