+++ src/main/java/io/magentys/donut/junit/model/Testsuites.java
@@ -1,0 +1,210 @@
+//
+// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.2.11 
+// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a> 
+// Any modifications to this file will be lost upon recompilation of the source schema. 
+// Generated on: 2017.09.22 at 01:07:35 PM AWST 
+//
+
+
+package io.magentys.donut.junit.model;
+
+import javax.xml.bind.annotation.*;
+import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
+
+
+/**
+ * <p>Java class for anonymous complex type.
+ * <p>
+ * <p>The following schema fragment specifies the expected content contained within this class.
+ * <p>
+ * <pre>
+ * &lt;complexType&gt;
+ *   &lt;complexContent&gt;
+ *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
+ *       &lt;sequence&gt;
+ *         &lt;element ref="{}testsuite" maxOccurs="unbounded" minOccurs="0"/&gt;
+ *       &lt;/sequence&gt;
+ *       &lt;attribute name="name" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
+ *       &lt;attribute name="time" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
+ *       &lt;attribute name="tests" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
+ *       &lt;attribute name="failures" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
+ *       &lt;attribute name="disabled" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
+ *       &lt;attribute name="errors" type="{http://www.w3.org/2001/XMLSchema}string" /&gt;
+ *     &lt;/restriction&gt;
+ *   &lt;/complexContent&gt;
+ * &lt;/complexType&gt;
+ * </pre>
+ */
+@XmlAccessorType(XmlAccessType.FIELD)
+@XmlType(name = "", propOrder = {
+        "testsuite"
+})
+@XmlRootElement(name = "testsuites")
+public class Testsuites
+        implements Serializable {
+
+    private final static long serialVersionUID = 1L;
+    protected List<Testsuite> testsuite;
+    @XmlAttribute(name = "name")
+    protected String name;
+    @XmlAttribute(name = "time")
+    protected String time;
+    @XmlAttribute(name = "tests")
+    protected String tests;
+    @XmlAttribute(name = "failures")
+    protected String failures;
+    @XmlAttribute(name = "disabled")
+    protected String disabled;
+    @XmlAttribute(name = "errors")
+    protected String errors;
+
+    /**
+     * Gets the value of the testsuite property.
+     * <p>
+     * <p>
+     * This accessor method returns a reference to the live list,
+     * not a snapshot. Therefore any modification you make to the
+     * returned list will be present inside the JAXB object.
+     * This is why there is not a <CODE>set</CODE> method for the testsuite property.
+     * <p>
+     * <p>
+     * For example, to add a new item, do as follows:
+     * <pre>
+     *    getTestsuite().add(newItem);
+     * </pre>
+     * <p>
+     * <p>
+     * <p>
+     * Objects of the following type(s) are allowed in the list
+     * {@link Testsuite }
+     */
+    public List<Testsuite> getTestsuite() {
+        if (testsuite == null) {
+            testsuite = new ArrayList<Testsuite>();
+        }
+        return this.testsuite;
+    }
+
+    /**
+     * Gets the value of the name property.
+     *
+     * @return possible object is
+     * {@link String }
+     */
+    public String getName() {
+        return name;
+    }
+
+    /**
+     * Sets the value of the name property.
+     *
+     * @param value allowed object is
+     *              {@link String }
+     */
+    public void setName(String value) {
+        this.name = value;
+    }
+
+    /**
+     * Gets the value of the time property.
+     *
+     * @return possible object is
+     * {@link String }
+     */
+    public String getTime() {
+        return time;
+    }
+
+    /**
+     * Sets the value of the time property.
+     *
+     * @param value allowed object is
+     *              {@link String }
+     */
+    public void setTime(String value) {
+        this.time = value;
+    }
+
+    /**
+     * Gets the value of the tests property.
+     *
+     * @return possible object is
+     * {@link String }
+     */
+    public String getTests() {
+        return tests;
+    }
+
+    /**
+     * Sets the value of the tests property.
+     *
+     * @param value allowed object is
+     *              {@link String }
+     */
+    public void setTests(String value) {
+        this.tests = value;
+    }
+
+    /**
+     * Gets the value of the failures property.
+     *
+     * @return possible object is
+     * {@link String }
+     */
+    public String getFailures() {
+        return failures;
+    }
+
+    /**
+     * Sets the value of the failures property.
+     *
+     * @param value allowed object is
+     *              {@link String }
+     */
+    public void setFailures(String value) {
+        this.failures = value;
+    }
+
+    /**
+     * Gets the value of the disabled property.
+     *
+     * @return possible object is
+     * {@link String }
+     */
+    public String getDisabled() {
+        return disabled;
+    }
+
+    /**
+     * Sets the value of the disabled property.
+     *
+     * @param value allowed object is
+     *              {@link String }
+     */
+    public void setDisabled(String value) {
+        this.disabled = value;
+    }
+
+    /**
+     * Gets the value of the errors property.
+     *
+     * @return possible object is
+     * {@link String }
+     */
+    public String getErrors() {
+        return errors;
+    }
+
+    /**
+     * Sets the value of the errors property.
+     *
+     * @param value allowed object is
+     *              {@link String }
+     */
+    public void setErrors(String value) {
+        this.errors = value;
+    }
+
+}
