+++ src/main/java/io/magentys/donut/reporter/EntityGsonBuilder.java
@@ -1,0 +1,12 @@
+package io.magentys.donut.reporter;
+
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+
+class EntityGsonBuilder {
+    static Gson createGson() {
+        return new GsonBuilder().setPrettyPrinting()
+                .create();
+    }
+}
